Switch range seekbar library
The old library is maintained, but has problems with accurately
selecting the top of the range when there are a lot of values (perhaps
b/c it's using floats under the hood).

The new library is simpler and unmaintained, but uses a Number generic
implementation, so it can be based around ints.